Reorganize GithubInfo componment
Rename GithubInfo and organize imports

change filename to lowercase

Move Post file back
