+++ src/web/src/components/Posts/GitHubInfo/Commits.tsx
@@ -1,6 +1,6 @@
-import { VscIssues } from 'react-icons/vsc';
+import { VscGitCommit } from 'react-icons/vsc';
 import { createStyles, makeStyles, Theme } from '@material-ui/core';
-import useGithubInfo from '../../hooks/use-githubInfo';
+import useGithubInfo from '../../../hooks/use-githubInfo';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -25,46 +25,49 @@ const useStyles = makeStyles((theme: Theme) =>
       marginRight: '1rem',
       verticalAlign: 'text-bottom',
     },
-    issues: {
+    commits: {
       paddingLeft: 0,
       display: 'flex',
       flexWrap: 'wrap',
       gap: '1.5rem',
     },
-    issue: {
+    commit: {
       listStyle: 'none',
     },
   })
 );
 
-const getIssueNumber = (issue: string) => issue.replace(/.+\/issues\/([0-9]+).*/, '$1');
+const SHORT_SHA_LENGTH = 7;
 
-const getIssueInfo = (issue: string) => {
-  const [, user, repo] = issue.split('/');
+const getCommitNumber = (url: string, length?: number) =>
+  url.replace(/.+\/(commit|commits)\/(\w{40}).*/, '$2').substr(0, length);
+
+const getCommitInfo = (commit: string) => {
+  const [, user, repo] = commit.split('/');
   return `${user}/${repo}`;
 };
 
-const Issues = () => {
+const Commits = () => {
   const classes = useStyles();
-  const { issues } = useGithubInfo();
+  const { commits } = useGithubInfo();
 
   return (
     <div className={classes.GitHubInfo}>
       <h2 className={classes.GitHubLinkTitle}>
-        <VscIssues className={classes.icon} />
-        {issues.length === 1 ? 'Issue' : 'Issues'}
+        <VscGitCommit className={classes.icon} />
+        {commits.length === 1 ? 'Commit' : 'Commits'}
       </h2>
-      <ul className={classes.issues}>
-        {issues.map((issue) => (
-          <li key={issue} className={classes.issue}>
+      <ul className={classes.commits}>
+        {commits.map((url) => (
+          <li key={url} className={classes.commit}>
             <a
-              href={`https://github.com${issue}`}
+              href={`https://github.com${url}`}
               target="_blank"
               rel="noreferrer"
-              title={`${getIssueInfo(issue)} Issue #${getIssueNumber(issue)}`}
+              title={`${getCommitInfo(url)} Commit ${getCommitNumber(url)}`}
               className={classes.link}
             >
-              #{getIssueNumber(issue)}
+              {getCommitNumber(url, SHORT_SHA_LENGTH)}
             </a>
           </li>
         ))}
@@ -73,4 +76,4 @@ const Issues = () => {
   );
 };
 
-export default Issues;
+export default Commits;
