+++ src/web/src/components/Posts/GitHubInfo/Issues.tsx
@@ -1,6 +1,6 @@
-import { VscGitCommit } from 'react-icons/vsc';
+import { VscIssues } from 'react-icons/vsc';
 import { createStyles, makeStyles, Theme } from '@material-ui/core';
-import useGithubInfo from '../../hooks/use-githubInfo';
+import useGithubInfo from '../../../hooks/use-githubInfo';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -25,49 +25,46 @@ const useStyles = makeStyles((theme: Theme) =>
       marginRight: '1rem',
       verticalAlign: 'text-bottom',
     },
-    commits: {
+    issues: {
       paddingLeft: 0,
       display: 'flex',
       flexWrap: 'wrap',
       gap: '1.5rem',
     },
-    commit: {
+    issue: {
       listStyle: 'none',
     },
   })
 );
 
-const SHORT_SHA_LENGTH = 7;
+const getIssueNumber = (issue: string) => issue.replace(/.+\/issues\/([0-9]+).*/, '$1');
 
-const getCommitNumber = (url: string, length?: number) =>
-  url.replace(/.+\/(commit|commits)\/(\w{40}).*/, '$2').substr(0, length);
-
-const getCommitInfo = (commit: string) => {
-  const [, user, repo] = commit.split('/');
+const getIssueInfo = (issue: string) => {
+  const [, user, repo] = issue.split('/');
   return `${user}/${repo}`;
 };
 
-const Commits = () => {
+const Issues = () => {
   const classes = useStyles();
-  const { commits } = useGithubInfo();
+  const { issues } = useGithubInfo();
 
   return (
     <div className={classes.GitHubInfo}>
       <h2 className={classes.GitHubLinkTitle}>
-        <VscGitCommit className={classes.icon} />
-        {commits.length === 1 ? 'Commit' : 'Commits'}
+        <VscIssues className={classes.icon} />
+        {issues.length === 1 ? 'Issue' : 'Issues'}
       </h2>
-      <ul className={classes.commits}>
-        {commits.map((url) => (
-          <li key={url} className={classes.commit}>
+      <ul className={classes.issues}>
+        {issues.map((issue) => (
+          <li key={issue} className={classes.issue}>
             <a
-              href={`https://github.com${url}`}
+              href={`https://github.com${issue}`}
               target="_blank"
               rel="noreferrer"
-              title={`${getCommitInfo(url)} Commit ${getCommitNumber(url)}`}
+              title={`${getIssueInfo(issue)} Issue #${getIssueNumber(issue)}`}
               className={classes.link}
             >
-              {getCommitNumber(url, SHORT_SHA_LENGTH)}
+              #{getIssueNumber(issue)}
             </a>
           </li>
         ))}
@@ -76,4 +73,4 @@ const Commits = () => {
   );
 };
 
-export default Commits;
+export default Issues;
