+++ src/web/src/components/Posts/GitHubInfo/PullRequests.tsx
@@ -1,6 +1,6 @@
 import { VscGitPullRequest } from 'react-icons/vsc';
 import { createStyles, makeStyles, Theme } from '@material-ui/core';
-import useGithubInfo from '../../hooks/use-githubInfo';
+import useGithubInfo from '../../../hooks/use-githubInfo';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
