+++ src/web/src/components/Posts/GitHubInfo/Users.tsx
@@ -1,7 +1,7 @@
 import { VscGithub } from 'react-icons/vsc';
 import { createStyles, makeStyles, Theme } from '@material-ui/core';
-import TelescopeAvatar from '../TelescopeAvatar';
-import useGithubInfo from '../../hooks/use-githubInfo';
+import TelescopeAvatar from '../../TelescopeAvatar';
+import useGithubInfo from '../../../hooks/use-githubInfo';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
